services/trade_producer/pyproject.toml
[tool.poetry]
name = "crypto_price_prediction"
version = "0.1.0"
description = "A real-time ML system that predicts short-term crypto prices"
authors = ["Diogo Matias <diogo.pomatias@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
quixstreams = "^2.8.0"
loguru = "^0.7.2"
pyyaml = "^6.0.1"
websockets = "^12.0"
pydantic-settings = "2.3.0"

[tool.poetry.dev-dependencies]
types-pyyaml = "^6.0.12.20240311"
pycln = "^2.4.0"
pytest = "^6.2.4"
black = "^21.7b0"
mypy = "^0.910"
pre-commit = "^2.13.0"
ruff = "^0.5.4"


[tool.ruff]
line-length = 80
select = [
    "E",  # Pycodestyle errors
    "F",  # Pyflakes errors
    "D",  # pydocstyle errors for docstring checks
    "B",  # Bugbear for finding common issues
    "I",  # isort for import sorting
]
ignore = [
    "D100",  # Ignore missing docstring in public module
    "D104",  # Ignore missing docstring in public package
]

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
